// web-application/src/types-openapi/api/models/ListMembershipRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Badminton Court Management System API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface ListMembershipRequest
 */
export interface ListMembershipRequest {
    /**
     * 
     * @type {number}
     * @memberof ListMembershipRequest
     */
    id?: number | null;
    /**
     * 
     * @type {string}
     * @memberof ListMembershipRequest
     */
    name?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListMembershipRequest
     */
    status?: string | null;
}

/**
 * Check if a given object implements the ListMembershipRequest interface.
 */
export function instanceOfListMembershipRequest(value: object): value is ListMembershipRequest {
    return true;
}

export function ListMembershipRequestFromJSON(json: any): ListMembershipRequest {
    return ListMembershipRequestFromJSONTyped(json, false);
}

export function ListMembershipRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): ListMembershipRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'id': json['id'] == null ? undefined : json['id'],
        'name': json['name'] == null ? undefined : json['name'],
        'status': json['status'] == null ? undefined : json['status'],
    };
}

export function ListMembershipRequestToJSON(json: any): ListMembershipRequest {
    return ListMembershipRequestToJSONTyped(json, false);
}

export function ListMembershipRequestToJSONTyped(value?: ListMembershipRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'id': value['id'],
        'name': value['name'],
        'status': value['status'],
    };
}