// web-application/src/types-openapi/api/models/ListMembershipResponse.ts
export interface ListMembershipResponse {
    /**
     * 
     * @type {Date}
     * @memberof ListMembershipResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof ListMembershipResponse
     */
    updatedAt?: Date | null;
    /**
     * 
     * @type {string}
     * @memberof ListMembershipResponse
     */
    createdBy?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListMembershipResponse
     */
    updatedBy?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListMembershipResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof ListMembershipResponse
     */
    name?: string | null;
    /**
     * 
     * @type {number}
     * @memberof ListMembershipResponse
     */
    price?: number;
    /**
     * 
     * @type {number}
     * @memberof ListMembershipResponse
     */
    discountPercent?: number;
    /**
     * 
     * @type {number}
     * @memberof ListMembershipResponse
     */
    durationDays?: number;
    /**
     * 
     * @type {string}
     * @memberof ListMembershipResponse
     */
    description?: string | null;
    /**
     * 
     * @type {string}
     * @memberof ListMembershipResponse
     */
    status?: string | null;
}