// web-application/src/types-openapi/api/models/UserMembershipResponse.ts
export interface UserMembershipResponse {
    /**
     * 
     * @type {number}
     * @memberof UserMembershipResponse
     */
    id?: number;
    /**
     * 
     * @type {number}
     * @memberof UserMembershipResponse
     */
    customerId?: number;
    /**
     * 
     * @type {number}
     * @memberof UserMembershipResponse
     */
    membershipId?: number;
    /**
     * 
     * @type {string}
     * @memberof UserMembershipResponse
     */
    membershipName?: string | null;
    /**
     * 
     * @type {Date}
     * @memberof UserMembershipResponse
     */
    startDate?: Date;
    /**
     * 
     * @type {Date}
     * @memberof UserMembershipResponse
     */
    endDate?: Date;
    /**
     * 
     * @type {boolean}
     * @memberof UserMembershipResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {CustomerDto}
     * @memberof UserMembershipResponse
     */
    customer?: CustomerDto;
    /**
     * 
     * @type {PaymentDto}
     * @memberof UserMembershipResponse
     */
    payment?: PaymentDto;
    /**
     * 
     * @type {Array<PaymentDto>}
     * @memberof UserMembershipResponse
     */
    payments?: Array<PaymentDto> | null;
    /**
     * 
     * @type {string}
     * @memberof UserMembershipResponse
     */
    status?: string | null;
}